BUG: Fix ReadAIMHeader to work with short images
The ReadAIMHeader function does not correctly parse AIM v020 and v030
headers, causing the unrecognized data type exception to be thrown.
More specifically, the block offsets slgihtly vary for AIM v020 and v030
and needed to be updated to correctly parse some header parameters such
as the image data type.